requirements.txt
numpy
spconv-cu118
cumm-cu118
nuscenes-devkit
tqdm
opencv-python
Pillow
ftfy

#pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

nvidia-cuda-runtime-cu11
nvidia-cuda-nvrtc-cu11

# 3d backbone
openmim

# mim install mmengine
# mim install mmcv==2.1.0
# mim install mmdet
# mim install mmsegmentation



